requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
diff-match-patch==20230430
Django==4.2.13
django-admin-interface==0.28.6
django-celery-beat==2.6.0
django-colorfield==0.11.0
django-filter==24.2
django-import-export==4.0.3
django-jalali-date==1.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
install==1.3.5
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.3.7
numpy==1.24.3
packaging==24.0
pandas==1.4.3
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
